+++ web/WEB-INF/metadata/migrate/20130114102543_create_table_reajusteindice.sql
@@ -2,8 +2,8 @@ CREATE TABLE reajusteindice (
     id bigint NOT NULL,
     indice_id bigint NOT NULL,
     tabelareajustecolaborador_id bigint NOT NULL,
-    valoratual double precision NOT NULL,
-    valorproposto double precision NOT NULL
+    valoratual numeric(8,2) NOT NULL,
+    valorproposto numeric(8,2) NOT NULL
 );--.go
 
 ALTER TABLE ONLY reajusteindice ADD CONSTRAINT reajusteindice_pkey PRIMARY KEY (id);--.go
